+++ AoC/Aoc/Year2019/OpCodeComputer/IInputChannel.cs
@@ -2,6 +2,6 @@
 {
 	public interface IInputChannel
 	{
-		int GetNext();
+		long GetNext();
 	}
 }
