+++ AoC/Aoc/Year2019/OpCodeComputer/IOutputChannel.cs
@@ -2,6 +2,6 @@
 {
 	public interface IOutputChannel
 	{
-		void Send(int val);
+		void Send(long val);
 	}
 }
